Refactor S1 DAG and plugins to use new configuration management
This commits replaces the existing `Sentinel1.py` DAG with a new `S1_GRD_1SDV.py` DAG that follows the naming conventions and concepts detailed in #118. It now uses the new configuration and `get_inputs_from` parameter between the task, except for the `sentinel1.gdal` subdag that remains unchanged. Further cleanup, refactoring is still tbd. In addition, it also improves existing plugins and adopts the S2 DAG to use the new features.

- The DHUS plugin now also check if an product exists before downloading and in that case it tries to download the next until the download limit is reached. Furthermore to `dhus_download_max` configuration has been reintroduced.

- The `S1MetadataOperator` and `ZipInspector` now allows to use the `get_inputs_from` parameter and returns it's output with `return`

- The `generate_wcs|wfs|wms_dict` and `publish_product` methods have now been moved to the `geoserver_plugin.py` so that it can be reused by other DAGs.

Resolves: #190